refactor record write code path.
This commit centralizes the record write state management into the record
write state machine, such that the functions called to convert of records
to blocks only mutates the block, and the record write state machine
manages mutation of the record state.  One benefit to this approach
is the ability to pass the record to more places as const.  This makes
reasoning about record modification straightforward.  A second benefit is
a net code reduction due to collapsing the write block header and write
block continuation header functions into a single and simpler function