Merged most changes (code cleanup) from patches received from Marco on 2015-10-22. Marco's patches are on branch marco/dedup-2015-10-22
Note that this doesn't merge the removal of the exablox device, since I still need that
for testing until Marco rolls the relevant changes into the standard file device.

The only other unmerged bit is a comment that doesn't make sense anymore,
which I had previously removed in my branch.